"Modified HTML structure of index page: changed 'col-md-5' class div to include 'mx-4' class"
This commit message describes the change in the code diff, which modifies the HTML structure of the index page. Specifically, it changes the 'col-md-5' class div to include the 'mx-4' class in the 'div' element. The message is concise and sticks to describing the changes in the code without making any interpretations.